requirements.txt
streamlit
requests
ollama
python-dotenv
sqlalchemy
bcrypt
Pillow
